Implement timing functions for linux

+++ include/system/utility.hh
@@ -2,11 +2,11 @@
 
 namespace System
 {
-	class Utility {
-	public:
-		virtual void Sleep(int milliseconds) = 0;
-		virtual int GetTicks() = 0;
-		
-		static Utility *GetInstance();
-	};
+    class Utility {
+    public:
+        virtual void Sleep(int milliseconds) = 0;
+        virtual unsigned long GetTicks() = 0;
+
+        static Utility *GetInstance();
+    };
 }
